+++ internal/channels/twilio/twilio.go
@@ -6,6 +6,7 @@ import (
 	"bytes"
 	"net/http"
 	"net/url"
+	"time"
 
 	"github.com/ajg/form"
 	"github.com/jaimeteb/chatto/internal/channels/messages"
@@ -36,9 +37,10 @@ type MessageIn struct {
 
 // Config models Twilio configuration
 type Config struct {
-	AccountSid string `mapstructure:"account_sid"`
-	AuthToken  string `mapstructure:"auth_token"`
-	Number     string `mapstructure:"number"`
+	AccountSid string        `mapstructure:"account_sid"`
+	AuthToken  string        `mapstructure:"auth_token"`
+	Number     string        `mapstructure:"number"`
+	Delay      time.Duration `mapstructure:"delay"`
 }
 
 // Client is the twilio client interface
@@ -51,6 +53,7 @@ type Channel struct {
 	Client Client
 	Number string
 	token  string
+	delay  time.Duration
 }
 
 // New returns an initialized telegram client
@@ -59,7 +62,7 @@ func New(config Config) *Channel {
 
 	log.Infof("Added Twilio client: %v", client.AccountSid)
 
-	return &Channel{Client: client.Messages, Number: config.Number, token: config.AuthToken}
+	return &Channel{Client: client.Messages, Number: config.Number, token: config.AuthToken, delay: config.Delay}
 }
 
 // SendMessage for Twilio
@@ -72,10 +75,14 @@ func (c *Channel) SendMessage(response *messages.Response) error {
 			imageURL = append(imageURL, u)
 		}
 
-		_, err := c.Client.SendMessage(c.Number, response.ReplyOpts.Twilio.Recipient, answer.Text, imageURL)
+		log.Debugf("Sending Twilio message: %+v", answer)
+		apiResp, err := c.Client.SendMessage(c.Number, response.ReplyOpts.Twilio.Recipient, answer.Text, imageURL)
 		if err != nil {
 			return err
 		}
+		log.Debugf("Twilio response: %+v", apiResp)
+
+		time.Sleep(c.delay)
 	}
 
 	return nil